# Implement basic common subexpression eliminate optimization