--- a/datafusion/src/execution/context.rs
+++ b/datafusion/src/execution/context.rs
@@ -24,9 +24,8 @@ use crate::{
     logical_plan::{PlanType, ToStringifiedPlan},
     optimizer::{
         aggregate_statistics::AggregateStatistics,
-        // common_subexpr_eliminate::CommonSubexprEliminate,
-        eliminate_limit::EliminateLimit,
-        hash_build_probe_order::HashBuildProbeOrder,
+        common_subexpr_eliminate::CommonSubexprEliminate,
+        eliminate_limit::EliminateLimit, hash_build_probe_order::HashBuildProbeOrder,
     },
     physical_optimizer::optimizer::PhysicalOptimizerRule,
 };
@@ -686,6 +685,7 @@ impl Default for ExecutionConfig {
             batch_size: 8192,
             optimizers: vec![
                 Arc::new(ConstantFolding::new()),
+                Arc::new(CommonSubexprEliminate::new()),
                 Arc::new(EliminateLimit::new()),
                 Arc::new(AggregateStatistics::new()),
                 Arc::new(ProjectionPushDown::new()),
@@ -693,7 +693,6 @@ impl Default for ExecutionConfig {
                 Arc::new(SimplifyExpressions::new()),
                 Arc::new(HashBuildProbeOrder::new()),
                 Arc::new(LimitPushDown::new()),
-                // Arc::new(CommonSubexprEliminate::new()),
             ],
             physical_optimizers: vec![
                 Arc::new(CoalesceBatches::new()),