--- a/datafusion/src/execution/context.rs
+++ b/datafusion/src/execution/context.rs
@@ -990,6 +990,7 @@ impl FunctionRegistry for ExecutionContextState {
 mod tests {
 
     use super::*;
+    use crate::logical_plan::{binary_expr, lit, Operator};
     use crate::physical_plan::functions::make_scalar_function;
     use crate::physical_plan::{collect, collect_partitioned};
     use crate::test;
@@ -2155,6 +2156,30 @@ mod tests {
             ];
             assert_batches_sorted_eq!(expected, &results);
         }
+    }
+
+    #[tokio::test]
+    async fn unprojected_filter() {
+        let mut ctx = ExecutionContext::new();
+        let df = ctx
+            .read_table(test::table_with_sequence(1, 3).unwrap())
+            .unwrap();
+
+        let df = df
+            .select(vec![binary_expr(col("i"), Operator::Plus, col("i"))])
+            .unwrap()
+            .filter(col("i").gt(lit(2)))
+            .unwrap();
+        let results = df.collect().await.unwrap();
+
+        let expected = vec![
+            "+--------------------------+",
+            "| ?table?.i Plus ?table?.i |",
+            "+--------------------------+",
+            "| 6                        |",
+            "+--------------------------+",
+        ];
+        assert_batches_sorted_eq!(expected, &results);
     }
 
     #[tokio::test]