--- a/datafusion/src/logical_plan/expr.rs
+++ b/datafusion/src/logical_plan/expr.rs
@@ -803,11 +803,12 @@ impl Expr {
     where
         R: ExprRewriter,
     {
-        match rewriter.pre_visit(&self)? {
+        let need_mutate = match rewriter.pre_visit(&self)? {
             RewriteRecursion::Mutate => return rewriter.mutate(self),
             RewriteRecursion::Stop => return Ok(self),
-            RewriteRecursion::Continue => {}
-        }
+            RewriteRecursion::Continue => true,
+            RewriteRecursion::Skip => false,
+        };
 
         // recurse into all sub expressions(and cover all expression types)
         let expr = match self {
@@ -923,7 +924,11 @@ impl Expr {
         };
 
         // now rewrite this expression itself
-        rewriter.mutate(expr)
+        if need_mutate {
+            rewriter.mutate(expr)
+        } else {
+            Ok(expr)
+        }
     }
 }
 
@@ -991,6 +996,8 @@ pub enum RewriteRecursion {
     Mutate,
     /// Do not rewrite / visit the children of this expression.
     Stop,
+    /// Keep recursive but skip mutate on this expression
+    Skip,
 }
 
 /// Trait for potentially recursively rewriting an [`Expr`] expression
@@ -1707,13 +1714,17 @@ fn create_name(e: &Expr, input_schema: &DFSchema) -> Result<String> {
         } => {
             let mut name = "CASE ".to_string();
             if let Some(e) = expr {
-                name += &format!("{:?} ", e);
+                let e = create_name(e, input_schema)?;
+                name += &format!("{} ", e);
             }
             for (w, t) in when_then_expr {
-                name += &format!("WHEN {:?} THEN {:?} ", w, t);
+                let when = create_name(w, input_schema)?;
+                let then = create_name(t, input_schema)?;
+                name += &format!("WHEN {} THEN {} ", when, then);
             }
             if let Some(e) = else_expr {
-                name += &format!("ELSE {:?} ", e);
+                let e = create_name(e, input_schema)?;
+                name += &format!("ELSE {} ", e);
             }
             name += "END";
             Ok(name)