@Dandandan 2021-07-29
Great @waynexia let me know when it's ready for review!

In the TCP-H benchmarks there are also a few that could benefit from CSE, so might make sense to inspect a few of those plans if it's implemented. Although it might not be affecting performance that much for those queries maybe.