@waynexia 2021-08-01
The implementation is almost finished. I plan to add more tests and make all cases pass next.

I tried to run the tpc-h benchmark and got a bad result :confused:

with cse:
>Running benchmarks with the following options: DataFusionBenchmarkOpt { query: 1, debug: false, iterations: 5, concurrency: 2, batch_size: 4096, path: "./data", file_format: "tbl", mem_table: false, partitions: 8 }
Query 1 iteration 0 took 2901.0 ms
Query 1 iteration 1 took 2767.9 ms
Query 1 iteration 2 took 3011.2 ms
Query 1 iteration 3 took 2861.0 ms
Query 1 iteration 4 took 2955.2 ms
Query 1 avg time: 2899.27 ms

master:
>Running benchmarks with the following options: DataFusionBenchmarkOpt { query: 1, debug: false, iterations: 5, concurrency: 2, batch_size: 4096, path: "./data", file_format: "tbl", mem_table: false, partitions: 8 }
Query 1 iteration 0 took 2751.7 ms
Query 1 iteration 1 took 2703.5 ms
Query 1 iteration 2 took 2787.9 ms
Query 1 iteration 3 took 2863.1 ms
Query 1 iteration 4 took 2831.3 ms
Query 1 avg time: 2787.50 ms