@waynexia 2021-08-18
Let me propose some relatively "big" changes:
- Make every result columns' name become the qualified name. (If we have any invariant on the column's name please let me know.)
- Add a new internal `Expr` for this case. Like `AliasWithExpr(Expr)` that can keep the original expression.
- Calculate the result table's header from query SQL. There might be some situations that are impossible to do this.

I haven't tried these, just put them into the discussion as some of them might be unrealistic.