@alamb 2021-08-19
Hi @waynexia  -- I plan to review your question/comments carefully tomorrow