@alamb 2021-08-20
Hi @waynexia  -- I noticed that the [failures in  `window_partition_by_order_by`](https://github.com/apache/arrow-datafusion/pull/792/checks?check_run_id=3252296013#step:4:1370) seem to be all related to window functions. Since window functions were added about the same time as we reworked the fully qualified names, I wonder if perhaps the bug you are chasing is something related to do with window functions specifically rather than a more general problem.

I think @houqp  might be the one with the most knowledge of aliasing / fully qualified column names at the moment.